Added hideTabs opt and fixed new files alignment
This should fix the issue related to new files being aligned to the left when on Zen mode, as pointed on defunkt/Zen#32. This was possible through the `onDidChangeActivePaneItem` method inside Atom's workspace. It's pretty simple actually.

Another change was to make the hiding tabs optional. By default Zen hides the tabs panel, but you can change this if you want. This is related to defunkt/Zen#35.