Add more options for the tab style while Zen is active
Added a dropdown menu to the settings which lets the user choose between
the 'hidden', 'single' and 'multiple' options.

The 'hidden' option is the default option and removes the whole tab bar
when Zen is active. The 'single' option displays only the tab of the
currently edited file and the 'multiple' option displays multiple tabs
at once.

Both the single and multiple options change the style of the open tabs
to keep them unobtrusive and zen-like. It removes all decorations, only
shows the title of the tabs and slightly highlights the currently active
one.